requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
timm
facenet-pytorch
matplotlib
streamlit
